Create annals-of-surgical-oncology-for-submission.csl
This is the style required by Annals of Surgical Oncology as defined by their submission instructions, found in the pdf on this page: http://www.springer.com/medicine/oncology/journal/10434 - look for the "Instructions for Authors" document to the right.

Note: It is _different_ from the actual journal style. The most notable difference is the italic DOI documents.